+++ bin/auto-encrypt-localhost.js
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-import AutoEncryptLocalhost from '../index.js'
-
-AutoEncryptLocalhost.https.createServer()