+++ test/index.js
@@ -5,6 +5,8 @@ import bent from 'bent'
 import test from 'tape'
 import AutoEncryptLocalhost from '../index.js'
 
+import '../bin/post-install.js'
+
 const downloadString = bent('GET', 'string')
 const downloadBuffer = bent('GET', 'buffer')
 
@@ -15,24 +17,19 @@ async function asyncForEach(array, callback) {
 }
 
 test('certificate creation', async t => {
-  t.plan(16)
+  const settingsPath = path.join(os.homedir(), '.small-tech.org', 'auto-encrypt-localhost')
 
-  const defaultSettingsPath = path.join(os.homedir(), '.small-tech.org', 'auto-encrypt-localhost')
-
-  const keyFilePath = path.join(defaultSettingsPath, 'localhost-key.pem')
-  const certFilePath = path.join(defaultSettingsPath, 'localhost.pem')
-
-  // Remove the settings path in case it already exists.
-  fs.removeSync(defaultSettingsPath)
+  const keyFilePath = path.join(settingsPath, 'localhost-key.pem')
+  const certFilePath = path.join(settingsPath, 'localhost.pem')
 
   // Run Auto Encrypt Localhost.
   const server = AutoEncryptLocalhost.https.createServer((request, response) => {
     response.end('ok')
   })
 
-  t.ok(fs.existsSync(path.join(defaultSettingsPath)), 'Main settings path exists.')
-  t.ok(fs.existsSync(path.join(defaultSettingsPath, 'rootCA.pem')), 'Local certificate authority exists.')
-  t.ok(fs.existsSync(path.join(defaultSettingsPath, 'rootCA-key.pem')), 'Local certificate authority private key exists.')
+  t.ok(fs.existsSync(path.join(settingsPath)), 'Main settings path exists.')
+  t.ok(fs.existsSync(path.join(settingsPath, 'rootCA.pem')), 'Local certificate authority exists.')
+  t.ok(fs.existsSync(path.join(settingsPath, 'rootCA-key.pem')), 'Local certificate authority private key exists.')
   t.ok(fs.existsSync(certFilePath), 'Local certificate exists.')
   t.ok(fs.existsSync(keyFilePath), 'Local certificate private key exists.')
 
@@ -75,26 +72,6 @@ test('certificate creation', async t => {
 
   // Wait the for the first server to close.
   await new Promise((resolve, reject) => { server.close(() => { resolve() }) })
-
-  //
-  // Custom settings path.
-  //
-
-  const customSettingsPath = path.join(os.homedir(), '.small-tech.org', 'auto-encrypt-localhost-custom-directory-test', 'second-level-directory')
-
-  // Remove the custom settings path in case it already exists.
-  fs.removeSync(customSettingsPath)
-
-  const server2 = AutoEncryptLocalhost.https.createServer({ settingsPath: customSettingsPath })
-
-  t.ok(fs.existsSync(path.join(customSettingsPath)), '(Custom settings path) Main directory exists.')
-  t.ok(fs.existsSync(path.join(customSettingsPath, 'rootCA.pem')), '(Custom settings path) Local certificate authority exists.')
-  t.ok(fs.existsSync(path.join(customSettingsPath, 'rootCA-key.pem')), '(Custom settings path) Local certificate authority private key exists.')
-  t.ok(fs.existsSync(path.join(customSettingsPath, 'localhost.pem')), '(Custom settings path) Local certificate exists.')
-  t.ok(fs.existsSync(path.join(customSettingsPath, 'localhost-key.pem')), '(Custom settings path) Local certificate private key exists.')
-
-  // Wait for the second server to close.
-  await new Promise((resolve, reject) => { server2.close(() => { resolve() }) })
 
   t.end()
 })